backend/requirements.txt
fastapi
uvicorn
openai
pillow
python-multipart
pymongo
python-dotenv
